+++ fuc/api/pybed.py
@@ -1,6 +1,6 @@
 """
 The pybed submodule is designed for working with BED files. It
-implements ``pybed.BedFrame`` which stores BED data as
+implements the ``pybed.BedFrame`` class which stores BED data as
 ``pandas.DataFrame`` via the `pyranges
 <https://github.com/biocore-ntnu/pyranges>`_ package to allow fast
 computation and easy manipulation. The submodule strictly adheres to the